refactor: use MaterialCommunityIcon in built-in components icons
