+++ src/core/Provider.tsx
@@ -1,10 +1,7 @@
 import * as React from 'react';
 import { ThemeProvider } from './theming';
-import {
-  Provider as SettingsProvider,
-  Settings,
-  defaultIcon,
-} from './settings';
+import { Provider as SettingsProvider, Settings } from './settings';
+import MaterialCommunityIcon from '../components/MaterialCommunityIcon';
 import PortalHost from '../components/Portal/PortalHost';
 import { Theme } from '../types';
 
@@ -18,7 +15,9 @@ export default class Provider extends React.Component<Props> {
   render() {
     return (
       <PortalHost>
-        <SettingsProvider value={this.props.settings || { icon: defaultIcon }}>
+        <SettingsProvider
+          value={this.props.settings || { icon: MaterialCommunityIcon }}
+        >
           <ThemeProvider theme={this.props.theme}>
             {this.props.children}
           </ThemeProvider>
